+++ test/test_sidebarx_widget.dart
@@ -6,14 +6,20 @@ class TestSidebarX extends StatelessWidget {
     Key? key,
     required this.controller,
     required this.items,
+    this.footerItems = const [],
     this.toggleButtonBuilder,
     this.animationDuration = const Duration(milliseconds: 300),
+    this.onExpansionChanged,
+    this.theme = const SidebarXTheme(),
   }) : super(key: key);
 
   final SidebarXController controller;
   final List<SidebarXItem> items;
+  final List<SidebarXItem> footerItems;
   final SidebarXBuilder? toggleButtonBuilder;
   final Duration animationDuration;
+  final void Function(SidebarXItem item, bool expanded)? onExpansionChanged;
+  final SidebarXTheme theme;
 
   @override
   Widget build(BuildContext context) {
@@ -24,7 +30,10 @@ class TestSidebarX extends StatelessWidget {
           animationDuration: animationDuration,
           controller: controller,
           items: items,
+          footerItems: footerItems,
           toggleButtonBuilder: toggleButtonBuilder,
+          onExpansionChanged: onExpansionChanged,
+          theme: theme,
         ),
       ),
     );